init json files if not exist

+++ scripts/preinstall.js
@@ -31,12 +31,26 @@ if (fs.existsSync(path.join(root, '.git'))) {
 ].forEach((loc) => {
   const src = path.join(__dirname, '..', loc);
   const dist = path.join(__dirname, '../dist', loc);
-  try {
-    fs.mkdirSync(path.dirname(src), { recursive: true });
-    fs.mkdirSync(path.dirname(dist), { recursive: true });
-    fs.writeFileSync(src, '{}');
-    fs.writeFileSync(dist, '{}');
-  } catch {
-    //
+  if (!fs.existsSync(dist)) {
+    tryCatch(() => {
+      fs.mkdirSync(path.dirname(dist), { recursive: true });
+      fs.writeFileSync(dist, '{}');
+    });
+  }
+  if (!fs.existsSync(src)) {
+    tryCatch(() => {
+      fs.mkdirSync(path.dirname(src), { recursive: true });
+      fs.writeFileSync(src, '{}');
+    });
   }
 });
+
+function tryCatch(fn) {
+  if (typeof fn === 'function') {
+    try {
+      fn();
+    } catch {
+      //
+    }
+  }
+}
